+++ pkg/httpclient/httpclient.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net/http"
 	"net/url"
-	"strings"
 	"time"
 
 	v4 "github.com/aws/aws-sdk-go-v2/aws/signer/v4"
@@ -161,7 +160,7 @@ func applyOAuthJWT(ctx context.Context, httpClient *http.Client, settings models
 		jwtConfig := jwt.Config{
 			Email:        settings.OAuth2Settings.Email,
 			TokenURL:     settings.OAuth2Settings.TokenURL,
-			PrivateKey:   []byte(strings.ReplaceAll(settings.OAuth2Settings.PrivateKey, "\\n", "\n")),
+			PrivateKey:   []byte(settings.OAuth2Settings.PrivateKey),
 			PrivateKeyID: settings.OAuth2Settings.PrivateKeyID,
 			Subject:      settings.OAuth2Settings.Subject,
 			Scopes:       []string{},